Add GL view.
