+++ app/build.gradle
@@ -22,7 +22,7 @@ android {
                 'APP_BUILD_SCRIPT=src/main/jni/Android.mk',
                 'NDK_APPLICATION_MK=src/main/jni/Application.mk'
     }
-    task ocvWrapper(type: Copy, description: 'Copy Java wrapper for OpenCV'){
+    task ocvWrapper(type: Copy, description: 'Copy Java wrapper for OpenCV') {
         from System.getenv("OCV_ANDROID") + '/native/libs'
         into 'src/main/jniLibs'
         include '**/*.so'
@@ -41,9 +41,9 @@ android {
 
 dependencies {
     compile fileTree(include: ['*.jar'], dir: 'libs')
-    compile 'com.android.support:appcompat-v7:25.3.1'
-    compile "com.android.support:support-v4:25.3.1"
-    compile "com.android.support:design:25.3.1"
     compile project(':openCVLibrary320')
     compile project(':cameraview')
+    compile 'com.android.support:appcompat-v7:25.3.1'
+    compile 'com.android.support:support-v4:25.3.1'
+    compile 'com.android.support:design:25.3.1'
 }
