+++ app/src/main/java/srijanparmeshwar/dolly/RenderView.java
@@ -1,0 +1,54 @@
+package srijanparmeshwar.dolly;
+
+import android.content.Context;
+import android.opengl.GLSurfaceView;
+import android.util.AttributeSet;
+
+import javax.microedition.khronos.egl.EGLConfig;
+import javax.microedition.khronos.opengles.GL10;
+
+/**
+ * Created by Srijan on 24/04/2017.
+ */
+
+public class RenderView extends GLSurfaceView {
+
+    private int width;
+    private int height;
+
+    public RenderView(Context context) {
+        super(context);
+        setup();
+    }
+
+    public RenderView(Context context, AttributeSet attrs) {
+        super(context, attrs);
+        setup();
+    }
+
+    private void setup() {
+        width = getMeasuredWidth();
+        height = getMeasuredHeight();
+        setEGLConfigChooser(8, 8, 8, 0, 16, 0);
+        setEGLContextClientVersion(3);
+        setRenderer(new Renderer());
+    }
+
+    private class Renderer implements GLSurfaceView.Renderer {
+
+        @Override
+        public void onSurfaceCreated(GL10 gl, EGLConfig config) {
+
+        }
+
+        @Override
+        public void onSurfaceChanged(GL10 gl, int width, int height) {
+
+        }
+
+        @Override
+        public void onDrawFrame(GL10 gl) {
+
+        }
+    }
+}
