+++ app/src/main/jni/Renderer.h
@@ -1,63 +1,46 @@
-#ifndef RENDERER_H
-#define RENDERER_H
-
-#include <opencv2/opencv.hpp>
-#include <opencv2/ximgproc/edge_filter.hpp>
-#include <opencv2/ximgproc/disparity_filter.hpp>
+#ifndef RENDERER2_H
+#define RENDERER2_H
 
 #include <string>
 #include <vector>
 
+#include <glm/vec3.hpp>
+#include <glm/mat4x4.hpp>
+#include <glm/gtc/matrix_transform.hpp>
+#include <glm/gtc/type_ptr.hpp>
+
+#include <GLES3/gl3.h>
+#include <EGL/egl.h>
+
 class Renderer {
-	// Two different views of a static scene.
-	cv::Mat A;
-	cv::Mat B;
+    int width;
+    int height;
 
-	// Rectification homographies.
-	cv::Mat H_A;
-	cv::Mat H_B;
+    float targetSize;
+    float targetDistance;
+    float z;
 
-	cv::Mat rectifiedA;
-	cv::Mat rectifiedB;
+    GLuint programID;
+    GLuint vertexShaderID;
+    GLuint fragmentShaderID;
 
-	cv::Mat depth;
+    GLint projectionID;
+    GLint zID;
 
-	struct Parameters {
-        float fps;
-        float length;
-		float targetDistance;
-		float targetSize;
-		float z_step;
-		float dz;
-	};
+    GLuint vertexBuffer;
+    GLuint colourBuffer;
 
-	Parameters parameters;
+    int N;
 
-	// Render novel views along given path.
-	/*void renderViews(
-			std::string filename, Renderer::PATH path,
-			float targetSize, float targetDistance, float fps, float videoLength
-		);*/
+    glm::mat4 projectionMatrix(float fov);
 
-	public:
-        enum PATH {
-            FORWARD,
-            BACKWARD
-        };
-
-		Renderer(cv::Mat frameA,
-                 cv::Mat frameB,
-                 float targetDistance,
-                 float targetSize,
-                 float fps,
-                 float length,
-                 PATH path);
-		~Renderer();
-		// Estimate depth using stereo matching.
-		void estimateDepth();
-		cv::Mat grabFrame();
-        void render(std::string path);
-		//void render(std::string path);
+    public:
+        Renderer(int w, int h, float s, float d) : width(w), height(w), targetSize(s), targetDistance(d), z(0.0f) {}
+        ~Renderer() {}
+        void init(std::string vertexShader, std::string fragmentShader);
+        void setMesh(std::vector<float> vertices, std::vector<float> colours);
+        void onSurfaceChanged(int w, int h);
+        void draw(float dz);
 };
 
 #endif
