+++ app/src/main/res/values/strings.xml
@@ -10,4 +10,5 @@
     <string name="storage_permission_confirmation">Storage permission granted</string>
     <string name="camera_permission_confirmation">Camera permission granted</string>
     <string name="preview_description">Preview of captured image.</string>
+    <string name="title_activity_render">RenderActivity</string>
 </resources>
