+++ app/src/main/res/values/styles.xml
@@ -20,4 +20,8 @@
         <item name="colorAccent">#0288D1</item>
     </style>
 
+    <style name="AppTheme.AppBarOverlay" parent="ThemeOverlay.AppCompat.Dark.ActionBar" />
+
+    <style name="AppTheme.PopupOverlay" parent="ThemeOverlay.AppCompat.Light" />
+
 </resources>
